fix(inspector): Work chips follow the active terminal again
The Inspector's branch and repo chips froze on whichever terminal painted
first and never moved again, while the working-directory line right beneath
them kept updating — the split a user sees as "the panel is half stale".

Introduced by #2037. It gated the chips on `Show when={git && colors()}`
and then RE-READ git inside the body:

    {(gitAndColors) => {
      const colors = gitAndColors();   // snapshot
      const git = props.meta.git!;     // snapshot

A non-keyed `<Show>` runs its body EXACTLY ONCE and keeps it mounted while
the condition stays truthy. Every terminal has git context, so the condition
never went falsy, the body never rebuilt, and both consts stayed pinned to
the first terminal for the life of the panel. `RightPanel` keeps both tabs
mounted and only `display:none`s the inactive one, so nothing else remounted
it either. The directory line was fine because it reads `props.meta.cwd`
inline in JSX.

Git facts and their paint now come from ONE `createMemo` behind ONE `Show`
accessor — the same glitch-free bundling `useTerminalStore.active()` uses
for `(id, meta)`. With a single value there is no second thing left to
snapshot, and a torn pair (this terminal's branch wearing the last one's
hue) is unrepresentable.

`WorkSection` moves to its own file: it is the one Inspector section with
real derivation rather than layout, and splitting it lets the regression be
covered by a render test without dragging the `../wire`-bound siblings
(Compose / Ports / Attach) into the module graph.

The new test switches the terminal and asserts the chips CHANGED — branch
text, repo text, worktree glyph, and both fleet hues. The existing e2e seam
only asserts `inspector-branch` is non-empty, which stale text satisfies;
non-emptiness was never the property that broke. Verified red-for-the-right-
reason before the fix: `expected 'resize-bug' to be 'docs/pesu-chat-rewrite'`.

Also from the same PR's audit: `data-testid="inspector-repo"` wrapped the
decorative monogram, so the chip's textContent read "Kkolu" rather than
"kolu". No step reads it today, so this was a latent trap rather than a
break; the testid moves onto the name span, the rule #2037 itself states for
the restore band's `repo-heading`.
