+++ website/src/content/changelog/unreleased.mdx
@@ -5,7 +5,7 @@ version: Unreleased
 ### [The Dock](/dock)
 
 - <Change kind="added" title="Every split has a row in the Dock" prs={[2046, 2056]}>Every split now appears as an indented entry beneath its parent terminal — including a plain shell, not only an agent. Click it and the keyboard lands in that pane; the parent stays highlighted as the active tile while the split highlights as the focused terminal. Every sub-entry carries the same status indicator as a top-level row — the shell identity glyph and activity motion for a plain shell, the agent brand mark and violet/amber attention wash when an agent is present — so shell and agent splits cannot drift. The old split-count chip is gone because the entries themselves are the inventory.</Change>
-- <Change kind="changed" title="Repo identity you can tell apart — on the dock and beyond" prs={[2035, 2037, 2038]}>With many repos open, the dock's section cues — a 3px spine, a 7px swatch, and a 9% pastel header wash — all dissolved into one pale list. Each repo is now a **card**: thicker spine, solid monogram tile (first letter of the repo), a stronger sticky header wash, and a little air between sections. That monogram is a **shared identity atom** — the same letter, fill formula, and fleet hue also lead the command palette (monogram · repo name · branch colour · spine), session-restore groups, the inspector's Work chips (repo and branch both match the dock), and the tile title bar (name stays the fleet pastel on the dark chrome, not app light-mode ink); the minimap tints each rect with a touch of repo colour. Attention (violet needs-you, amber unread) stays the loudest interrupt; host palette headers pick up a soft host-hue wash without competing.</Change>
+- <Change kind="changed" title="Repo identity you can tell apart — on the dock and beyond" prs={[2035, 2037, 2038]}>With many repos open, the dock's section cues — a 3px spine, a 7px swatch, and a 9% pastel header wash — all dissolved into one pale list. Each repo is now a **card**: thicker spine, solid monogram tile (first letter of the repo), a stronger sticky header wash, and a little air between sections. That monogram is a **shared identity atom** — the same letter, fill formula, and fleet hue also lead the command palette (monogram · repo name · branch colour · spine), session-restore groups, the inspector's Work chips (repo and branch both match the dock, and both repaint as you move between terminals), and the tile title bar (name stays the fleet pastel on the dark chrome, not app light-mode ink); the minimap tints each rect with a touch of repo colour. Attention (violet needs-you, amber unread) stays the loudest interrupt; host palette headers pick up a soft host-hue wash without competing.</Change>
 
 ### [The Canvas](/canvas)
